PSP-9258 prevent L&L files from zooming out too far when toggling fullscreen.
