+++ source/frontend/src/features/mapSideBar/acquisition/models/LeaseStakeholderModel.ts
@@ -69,7 +69,6 @@ export class CompReqLeaseStakeholderModel {
   }
 
   public toApi(): ApiGen_Concepts_CompReqLeaseStakeholder {
-    debugger;
     const compReqPayeeModel: ApiGen_Concepts_CompReqLeaseStakeholder = {
       ...getEmptyBaseAudit(),
       compReqLeaseStakeholderId: this.apiId,
